Correctly calculate marker positioning at n*100% zoom.
This was only coincidental before. The new calculation is parameterized
off marker width (here, 7px + 2*2px border), and stone width (here 28px
+ 2*1px border). It was broken before at 200% browser zoom because it
was off by 1 pixel to the top-left, which made no sense because we
shouldn't have had fractional pixels. The reason is that (11+1)/2 + 1
happened to be 7, which was 28/4, but we were adding 1px as a kludge
factor, which then doubled to 2px at 200% zoom, causing it to break.