Make sure that an undo correctly unsets existing game state.
The broken behavior was introduced in
bb98f57e007ec3c49cad556c9d3bed167e3bf7c8 as part of a refactor. The
problem is that if we render the previous move (after popping), we're
supposed to reset all of the game state. We weren't doing that --
instead, we were resetting the state of the points we were supposed to
be using as the authority.

One underlying issue here is that the copy of the points stored in
`moves` should probably be immutable. Another is that we're maintaining
a copy `moves[n].points` in `intersections[y][x]`, which is inherently
going to suffer cache woes.