Stop storing the "A19"-style coordinates of each move.
These coordinates are specific to a particular representation of the
board, and shouldn't form a crucial part of the game state. Anything
manipulating the game state should have enough information already to
work with a Game instance and do its thing. (The game state is also not
"public" in the sense of it being a representation that you would export
and import in some fashion.)

The coordinate values on the game state object were entirely unused.