Make sure to count only unique captured stones.
If a group of stones are captured in such a way that their last taken
liberty is a _shared_ liberty, then we mistakenly count points twice.
The reason is related to the fact that the last taken liberty has 2+
neighbors which are taken, and we "look" through both of them to
determine the group.