Stop temporarily adding a game move just to calculate ko points.
By shifting `inAtari` to be on a `GameState`, we can now compute a game
state with the played stone, then calculate calculate liberties off the
resulting state, without saving the game state anywhere. The generated
state used in this calculation is technically not complete, but we only
need it to be correct for the purposes of liberty calculations.