Remove an unnecessary capturing check.
In an early version of the suicide check, captures were not accounted
for, it was a purely liberty-based condition. However it currently does
account for captures: a move that is a liberty-based suicide but that
captures will set `isSuicide` to `false`, so the check for `isCapturing`
is unnecessary.

As it happens, `isCapturing` was also broken. For any non-empty point,
it was always `false`! It was looking at `isEmpty`, not `isEmpty()`, and
`!neighbor.isEmpty` is thus `false`, which means there are always `0`
captured neighbors. So all non-empty points were considered to be
non-capturing.

Ordinarily, this would've caused problems. The check was:

  isSuicide && !isCapturing

If `isSuicide` were purely liberty-based and there were captures, then
the _broken_ `isCapturing` value would've resulted in:

  true && !false == true && true && true

This would've made the move illegal when it shouldn't have been. But
since `isSuicide` does indeed account for captures, in the case where
there are captures, `isSuicide` is false and so the second condition of
`!isCapturing` is never checked.