Restructure and refactor a bunch of things.
There's now a looser coupling between a game and a DOM renderer,
primarily as part of trying to get a structure that will work for a
client playing against a server.

The specific notion of an equivalence scoring strategy is broken apart
into a specific flag indicating whether the scorer uses pass stones and
whether the game must end by white passing last. The user interface of
setting the scoring type to "equivalence" is the same, it's just now
driving two separate flags.