Add a Client intended for playing as a single colour against a server.
With this commit, it is possible to play against a server from start to
finish in the same way as you would locally using just a Game object.
There are some loose assumptions about the interface a server must
provide, and they may be wrong. This is just an initial first pass.

The Client involves 3 hooks:

  * submitPlay, for submitting moves to a server.
  * submitPass, for submitting a pass to a server.
  * submitMarkDeadAt, for submitting stones that have been marked dead.

A Client wraps a Game object. If the submission was successful based on
hook callbacks, the action is executed locally on that Game object.

An assumption in this implementation is that the local client is
reliably configured to match the server and can be "trusted" with
non-critical responsibilities. For example, the local client can
calculate illegal moves (to avoid unnecessary calls to the server), and
can display calculated territory points based on the scoring rule in
used.

Critical responsibilities -- for example, anything that would let
you cheat -- is assumed to rest with the server. A full application
would _not_ rely on this library to locally calculate the final score of
the game. A discrepancy between, say, the displayed scored points of
territory and the actual computed server score indicates either a
local rendering bug, a misconfiguring between the client and server, or
a server bug in calculating the score.

Since one of the non-critical responsibilities managed locally is move
legality, and since move legality is dependent on move history, the
Client simply relies on a Game to take care of things, instead of
duplicating the Game object's implementation.

Note that this initial implementation is currently unfit for use cases
like post-game review. In those scenarios, a full application would
likely have move history and board position states given by the server.
Instead of cramming that stuff into a Game object, you would navigate a
separate tree and dump the state for a given node into a board
representation. This commit does not aim to address the transition
between a Game and some kind of review phase, nor does it aim to
implement the review phase at all.

Any features not in Tenuki will inherently be unsupported by the Client
object. Naturally that will change as more features are added to Game,
so that it can be configured correctly locally to match features that a
server has.