Add fuzzy stone placement as a configurable option.
Setting this flag to true automatically sets the stones to be textured.

This also technically makes it possible for textured stones to be 1
pixel smaller in radius, even when fuzzy stone placement is off.

Note that even when the stones are smaller, there are no gaps between
intersections because of the 1px transparent border. This is important,
because if there are gaps you get two bad outcomes:

  1. With a pointer device, hovering will flicker as you move between
     intersections.
  2. Much worse: on a touch device, tapping may tap between
     intersections, resulting in what in effectively a "dead" tap.