Recalculate board dimensions and avoid pinch zooming issues.
As part of this change, a base `Renderer` class is split out of
`DOMRenderer`. The relevant work for computing resizing is in
`computeSizing`. It fell out of extraction work for an SVG renderer.

There are functional changes:

  1. When zoomed in, you can't pinch zoom in, and trying to pinch zoom
     out will just zoom you out as if you'd tapped cancel. This is to
     stop double-zooming issues.
  2. When the browser window resizes, the board's dimensions are
     recalculated and updated.

Board size recalculation lets elements with a relative size auto-shrink
if the window is made smaller. It also means that rotating a phone
between landscape and portrait will resize the board, whereas before it
was necessary to fully refresh the page.

One consequence of this is that -- like the examples in `examples/` --
it's possible for the board to be limited in size beyond a certain
point. That'll happen if `.tenuki-board` is set to be a relative size.
If it's an absolute size then there won't be such a restriction.