Don't treat handicap stones as initially legal moves for white.
With non-free-placement hanidicap stones, `_moves` is empty until
white's first move (because we're treating the board setup with the
handicap stones as not-a-move), and so an intersection with a handicap
stone on it is mistakenly treated as a legal move for white. White could
in fact play at the occupied point!

This was visually showing up as a bug on the board where hovering over
a handicap stone would remove the stone on `mouseenter` because of the
way `hoverValue` checks for move legality.