Auto-scale the board up as well as down.
This adds some logic to force a re-raster when recalculating the board's
dimensions. The re-raster issue was there before when scaling down, but
when scaling up it's more obvious and needs fixing.

The change to `overflow` being moved to the top-level container is to
avoid issues where the board's original size was still being used in
certain cases. For example, if the board was set to 100% of the viewport
height/width, then even when the board was smaller than the full height
of the viewport (in a narrow portrait orientation), the browser would
decide it needed ~20 pixels of space for a vertical scrollbar even
though no such scrollbar was visible. The CSS transform scale() made the
board smaller, but the browser was reducing the document size by ~20px
to account for scrollbars that aren't actually there. After adding
`overflow: hidden`, the browser uses the effective post-scale() size of
the board for real.