Remove bower.json.
Bower's idea of a "release" is a Git tag in a repo containing source
code. Since this inherently requires the anti-pattern of including
built/generated files in Git, I'm removing support for Bower in favour
of being able to remove generated files from the repo.

This seems to be a long-standing issue with Bower. See, for example,
Bower GitHub issue 455.