End the game after 2 passes even under equivalence scoring rules
The previous implementation of equivalence scoring was incorrect. 3
passes are not required to end the game if the 2nd pass is by black.

Instead, the end of the game is _signaled_ by 2 consecutive passes, and
an agreement phase begins. After that agreement phase is over, then
white may be required to hand over one final pass stone. If there are
any disputes, then white's 3rd and final "pass" never happens.

This commit stops adding a 3rd consecutive white pass to the board state
and moves the "final white pass" logic into the scoring step.