Fix some region checking code for intersections on the board edge
A previous change made intersectionAt throw an error when given (y, x)
values outside the board. This caused a problem with Region checking for
square four and curved four shapes, which are only sometimes run
depending on the region's boundary length. Those shape-checking
functions assume that intersectionAt() will actually return undefined
values. For now, those two functions explicitly check if the given
coordinates are on the board.

This also fixes a silly oversight with an error message containing
hard-coded y and x values.